автоматически создается чат для клиента
